Complete ansible module

+++ app/jobs/machine_delete_job.rb
@@ -4,8 +4,17 @@ class MachineDeleteJob < Struct.new(:machine_id)
     machine = Machine.find machine_id
     if machine
       machine.stop
+      while machine.show_machine == CloudToolkit::STATUS_ONPROCESS
+        sleep(5.seconds)
+      end
+      if machine.show_machine == CloudToolkit::STATUS_ERROR
+        fail
+      end
       machine.destroy
     end
   end
 
+  def max_attempts
+    3
+  end
 end
